Add remove order operation

+++ src/main/java/cz/cvut/fit/ryntluka/service/CustomerService.java
@@ -1,39 +1,43 @@
 package cz.cvut.fit.ryntluka.service;
 
 import cz.cvut.fit.ryntluka.dto.CustomerCreateDTO;
-import cz.cvut.fit.ryntluka.dto.CustomerDTO;
+import cz.cvut.fit.ryntluka.dto.PlanetDTO;
 import cz.cvut.fit.ryntluka.entity.Customer;
+import cz.cvut.fit.ryntluka.entity.Planet;
 import cz.cvut.fit.ryntluka.exceptions.EntityMissingException;
 import cz.cvut.fit.ryntluka.repository.CustomerRepository;
+import cz.cvut.fit.ryntluka.repository.PlanetRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class CustomerService {
 
     private final CustomerRepository customerRepository;
+    private final PlanetRepository planetRepository;
 
     @Autowired
-    public CustomerService(CustomerRepository customerRepository) {
+    public CustomerService(CustomerRepository customerRepository, PlanetRepository planetRepository) {
         this.customerRepository = customerRepository;
+        this.planetRepository = planetRepository;
     }
 
     /*================================================================================================================*/
 
     @Transactional
-    public Customer create(CustomerCreateDTO customerCreateDTO) {
+    public Customer create(CustomerCreateDTO customerCreateDTO) throws EntityMissingException {
         return customerRepository.save(
                 new Customer(
                         customerCreateDTO.getFirstName(),
                         customerCreateDTO.getLastName(),
-                        customerCreateDTO.getEmail())
+                        customerCreateDTO.getEmail(),
+                        planetRepository.findById(customerCreateDTO.getPlanetId())
+                                .orElseThrow(EntityMissingException::new))
         );
     }
 
@@ -43,8 +47,8 @@ public class CustomerService {
         return customerRepository.findAll();
     }
 
-    public Optional<Customer> findByLastName (String lastName) {
-        return customerRepository.findByLastName(lastName);
+    public List<Customer> findAllByLastName(String lastName) {
+        return customerRepository.findAllByLastName(lastName);
     }
 
     public Optional<Customer> findById(int id) {
@@ -67,13 +71,17 @@ public class CustomerService {
         customer.setFirstName(customerCreateDTO.getFirstName());
         customer.setLastName(customerCreateDTO.getLastName());
         customer.setEmail(customerCreateDTO.getEmail());
+        customer.setPlanet(planetRepository.findById(customerCreateDTO.getPlanetId())
+                .orElseThrow(EntityMissingException::new));
         return customer;
     }
 
     /*================================================================================================================*/
 
     @Transactional
-    public void delete(int id) {
+    public void delete(int id) throws EntityMissingException {
+        if (customerRepository.findById(id).isEmpty())
+            throw new EntityMissingException();
         customerRepository.deleteById(id);
     }
 }
